feat: move inner classes to the main package
