+++ src/main/java/com/kjh/wms/product/feature/RegisterProduct.java
@@ -1,5 +1,7 @@
 package com.kjh.wms.product.feature;
 
+import com.kjh.wms.product.domain.*;
+
 import static org.springframework.util.Assert.hasText;
 import static org.springframework.util.Assert.notNull;
