+++ src/test/java/com/kjh/wms/product/feature/RegisterProductTest.java
@@ -1,5 +1,8 @@
 package com.kjh.wms.product.feature;
 
+import com.kjh.wms.product.domain.Category;
+import com.kjh.wms.product.domain.ProductRepository;
+import com.kjh.wms.product.domain.TemperatureZone;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
